HACK: Add toThrowAsync
A quick hack to explore possible solutions to issue #353. I think a
better solution would be to have the AsyncTest somehow determine which
matchers should perform an "await". With this solution I'm forcing
the programmer to pick a matcher, toThrowAsync, which does an await.

I added a couple tests to test/unit-tests/expect-tests/to-throw.spec.ts.