Fix for test-runner `is used before being assigned`
Without this change TSC is correctly giving warnings that
`currentTestFixtureResults` and `currentTestResults` are
`used before being assigned`.

Also, with this change but not using stricter compiler option test coverage
is missing for line 89, which means `currentTestResults` is always valid.
But if I put that `then` portion of `if (currentTestResults) {` into
`if (currentTestFixtureResults && (!previousTestItem || previousTestItem.test !== testItem.test)) {`
`_runTests` will fail.

It seems this loop is a little odd and maybe needs some refractoring, but
I'm not sure exactly how it works so I need some guidance.